修改连锁类型的bug  将校验移动到controller中  去掉service中的业务，移动到business

+++ xpaas-channel-center-channel/src/main/java/com/deepexi/channel/controller/ChainController.java
@@ -93,6 +93,10 @@ public class ChainController {
     @ApiOperation(value = "根据id批量删除连锁", notes = "根据id删除连锁")
     public Payload<Boolean> delete(@PathVariable(value = "id") String id) {
         List<Long> ids = Arrays.stream(id.split(",")).map(Long::parseLong).collect(Collectors.toList());
-        return new Payload<>(chainBusinessService.deleteChain(ids));
+        if(chainBusinessService.deleteVerification(ids)){
+            return new Payload<>(chainBusinessService.deleteChain(ids));
+        }else{
+            return new Payload<>(false);
+        }
     }
 }
